organized project dir
